Replace use of get_file() with download_file()
The Yahoo Groups API is a bit flakey and often returns 400 errors. The
get_file() method isn't robust enough to handle these 400 errors and
will cause the script to exit immediately. The download_file() method
however is able to handle retrying to get files, and with some minor
modifications can be a drop-in replacement for get_file(), which should
usually allow the script to complete.